Refactor: libcrmcommon: separate generic string functions into own source file
utils.c was a bit large
