+++ include/crm_internal.h
@@ -146,14 +146,6 @@ extern int node_score_yellow;
 extern int node_score_infinity;
 
 /* Assorted convenience functions */
-static inline int
-crm_strlen_zero(const char *s)
-{
-    return !s || *s == '\0';
-}
-
-char *add_list_element(char *list, const char *value);
-
 int crm_pid_active(long pid, const char *daemon);
 void crm_make_daemon(const char *name, gboolean daemonize, const char *pidfile);
 
@@ -198,14 +190,7 @@ crm_set_bit(const char *function, const char *target, long long word, long long 
 #  define set_bit(word, bit) word = crm_set_bit(__FUNCTION__, NULL, word, bit)
 #  define clear_bit(word, bit) word = crm_clear_bit(__FUNCTION__, NULL, word, bit)
 
-void g_hash_destroy_str(gpointer data);
-
-long long crm_int_helper(const char *text, char **end_text);
-char *crm_concat(const char *prefix, const char *suffix, char join);
 char *generate_hash_key(const char *crm_msg_reference, const char *sys);
-
-bool crm_compress_string(const char *data, int length, int max, char **result,
-                         unsigned int *result_len);
 
 /*! remote tcp/tls helper functions */
 typedef struct crm_remote_s crm_remote_t;
