+++ lib/common/utils.c
@@ -428,24 +428,6 @@ verify_all_options(GHashTable * options, pe_cluster_option * option_list, int le
 }
 
 char *
-crm_concat(const char *prefix, const char *suffix, char join)
-{
-    int len = 0;
-    char *new_str = NULL;
-
-    CRM_ASSERT(prefix != NULL);
-    CRM_ASSERT(suffix != NULL);
-    len = strlen(prefix) + strlen(suffix) + 2;
-
-    new_str = malloc(len);
-    if(new_str) {
-        sprintf(new_str, "%s%c%s", prefix, join, suffix);
-        new_str[len - 1] = 0;
-    }
-    return new_str;
-}
-
-char *
 generate_hash_key(const char *crm_msg_reference, const char *sys)
 {
     char *hash_key = crm_concat(sys ? sys : "none", crm_msg_reference, '_');
@@ -454,30 +436,6 @@ generate_hash_key(const char *crm_msg_reference, const char *sys)
     return hash_key;
 }
 
-
-char *
-crm_itoa_stack(int an_int, char *buffer, size_t len)
-{
-    if (buffer != NULL) {
-        snprintf(buffer, len, "%d", an_int);
-    }
-
-    return buffer;
-}
-
-char *
-crm_itoa(int an_int)
-{
-    int len = 32;
-    char *buffer = NULL;
-
-    buffer = malloc(len + 1);
-    if (buffer != NULL) {
-        snprintf(buffer, len, "%d", an_int);
-    }
-
-    return buffer;
-}
 
 int
 crm_user_lookup(const char *name, uid_t * uid, gid_t * gid)
@@ -610,134 +568,6 @@ compare_version(const char *version1, const char *version2)
 }
 
 gboolean do_stderr = FALSE;
-
-void
-g_hash_destroy_str(gpointer data)
-{
-    free(data);
-}
-
-#include <sys/types.h>
-/* #include <stdlib.h> */
-/* #include <limits.h> */
-
-long long
-crm_int_helper(const char *text, char **end_text)
-{
-    long long result = -1;
-    char *local_end_text = NULL;
-    int saved_errno = 0;
-
-    errno = 0;
-
-    if (text != NULL) {
-#ifdef ANSI_ONLY
-        if (end_text != NULL) {
-            result = strtol(text, end_text, 10);
-        } else {
-            result = strtol(text, &local_end_text, 10);
-        }
-#else
-        if (end_text != NULL) {
-            result = strtoll(text, end_text, 10);
-        } else {
-            result = strtoll(text, &local_end_text, 10);
-        }
-#endif
-
-        saved_errno = errno;
-/* 		CRM_CHECK(errno != EINVAL); */
-        if (errno == EINVAL) {
-            crm_err("Conversion of %s failed", text);
-            result = -1;
-
-        } else if (errno == ERANGE) {
-            crm_err("Conversion of %s was clipped: %lld", text, result);
-
-        } else if (errno != 0) {
-            crm_perror(LOG_ERR, "Conversion of %s failed:", text);
-        }
-
-        if (local_end_text != NULL && local_end_text[0] != '\0') {
-            crm_err("Characters left over after parsing '%s': '%s'", text, local_end_text);
-        }
-
-        errno = saved_errno;
-    }
-    return result;
-}
-
-int
-crm_parse_int(const char *text, const char *default_text)
-{
-    int atoi_result = -1;
-
-    if (text != NULL) {
-        atoi_result = crm_int_helper(text, NULL);
-        if (errno == 0) {
-            return atoi_result;
-        }
-    }
-
-    if (default_text != NULL) {
-        atoi_result = crm_int_helper(default_text, NULL);
-        if (errno == 0) {
-            return atoi_result;
-        }
-
-    } else {
-        crm_err("No default conversion value supplied");
-    }
-
-    return -1;
-}
-
-gboolean
-safe_str_neq(const char *a, const char *b)
-{
-    if (a == b) {
-        return FALSE;
-
-    } else if (a == NULL || b == NULL) {
-        return TRUE;
-
-    } else if (strcasecmp(a, b) == 0) {
-        return FALSE;
-    }
-    return TRUE;
-}
-
-gboolean
-crm_is_true(const char *s)
-{
-    gboolean ret = FALSE;
-
-    if (s != NULL) {
-        crm_str_to_boolean(s, &ret);
-    }
-    return ret;
-}
-
-int
-crm_str_to_boolean(const char *s, int *ret)
-{
-    if (s == NULL) {
-        return -1;
-
-    } else if (strcasecmp(s, "true") == 0
-               || strcasecmp(s, "on") == 0
-               || strcasecmp(s, "yes") == 0 || strcasecmp(s, "y") == 0 || strcasecmp(s, "1") == 0) {
-        *ret = TRUE;
-        return 1;
-
-    } else if (strcasecmp(s, "false") == 0
-               || strcasecmp(s, "off") == 0
-               || strcasecmp(s, "no") == 0 || strcasecmp(s, "n") == 0 || strcasecmp(s, "0") == 0) {
-        *ret = FALSE;
-        return 1;
-    }
-    return -1;
-}
 
 #ifndef NUMCHARS
 #  define	NUMCHARS	"0123456789."
@@ -1419,42 +1249,6 @@ crm_make_daemon(const char *name, gboolean daemonize, const char *pidfile)
     (void)open(devnull, O_WRONLY);      /* Stdout: fd 1 */
     close(STDERR_FILENO);
     (void)open(devnull, O_WRONLY);      /* Stderr: fd 2 */
-}
-
-char *
-crm_strip_trailing_newline(char *str)
-{
-    int len;
-
-    if (str == NULL) {
-        return str;
-    }
-
-    for (len = strlen(str) - 1; len >= 0 && str[len] == '\n'; len--) {
-        str[len] = '\0';
-    }
-
-    return str;
-}
-
-gboolean
-crm_str_eq(const char *a, const char *b, gboolean use_case)
-{
-    if (use_case) {
-        return g_strcmp0(a, b) == 0;
-
-        /* TODO - Figure out which calls, if any, really need to be case independent */
-    } else if (a == b) {
-        return TRUE;
-
-    } else if (a == NULL || b == NULL) {
-        /* shouldn't be comparing NULLs */
-        return FALSE;
-
-    } else if (strcasecmp(a, b) == 0) {
-        return TRUE;
-    }
-    return FALSE;
 }
 
 char *
@@ -2260,43 +2054,6 @@ determine_request_user(const char *user, xmlNode * request, const char *field)
 }
 #endif
 
-/*
- * This re-implements g_str_hash as it was prior to glib2-2.28:
- *
- *   http://git.gnome.org/browse/glib/commit/?id=354d655ba8a54b754cb5a3efb42767327775696c
- *
- * Note that the new g_str_hash is presumably a *better* hash (it's actually
- * a correct implementation of DJB's hash), but we need to preserve existing
- * behaviour, because the hash key ultimately determines the "sort" order
- * when iterating through GHashTables, which affects allocation of scores to
- * clone instances when iterating through rsc->allowed_nodes.  It (somehow)
- * also appears to have some minor impact on the ordering of a few
- * pseudo_event IDs in the transition graph.
- */
-guint
-g_str_hash_traditional(gconstpointer v)
-{
-    const signed char *p;
-    guint32 h = 0;
-
-    for (p = v; *p != '\0'; p++)
-        h = (h << 5) - h + *p;
-
-    return h;
-}
-
-guint
-crm_strcase_hash(gconstpointer v)
-{
-    const signed char *p;
-    guint32 h = 0;
-
-    for (p = v; *p != '\0'; p++)
-        h = (h << 5) - h + g_ascii_tolower(*p);
-
-    return h;
-}
-
 void *
 find_library_function(void **handle, const char *lib, const char *fn, gboolean fatal)
 {
@@ -2325,25 +2082,6 @@ find_library_function(void **handle, const char *lib, const char *fn, gboolean f
     }
 
     return a_function;
-}
-
-char *
-add_list_element(char *list, const char *value)
-{
-    int len = 0;
-    int last = 0;
-
-    if (value == NULL) {
-        return list;
-    }
-    if (list) {
-        last = strlen(list);
-    }
-    len = last + 2;             /* +1 space, +1 EOS */
-    len += strlen(value);
-    list = realloc_safe(list, len);
-    sprintf(list + last, " %s", value);
-    return list;
 }
 
 void *
@@ -2398,57 +2136,6 @@ crm_md5sum(const char *buffer)
     return digest;
 }
 
-#include <time.h>
-#include <bzlib.h>
-
-bool
-crm_compress_string(const char *data, int length, int max, char **result, unsigned int *result_len)
-{
-    int rc;
-    char *compressed = NULL;
-    char *uncompressed = strdup(data);
-    struct timespec after_t;
-    struct timespec before_t;
-
-    if(max == 0) {
-        max = (length * 1.1) + 600; /* recomended size */
-    }
-
-#ifdef CLOCK_MONOTONIC
-    clock_gettime(CLOCK_MONOTONIC, &before_t);
-#endif
-
-    /* coverity[returned_null] Ignore */
-    compressed = malloc(max);
-
-    *result_len = max;
-    rc = BZ2_bzBuffToBuffCompress(compressed, result_len, uncompressed, length, CRM_BZ2_BLOCKS, 0,
-                                  CRM_BZ2_WORK);
-
-    free(uncompressed);
-
-    if (rc != BZ_OK) {
-        crm_err("Compression of %d bytes failed: %s (%d)", length, bz2_strerror(rc), rc);
-        free(compressed);
-        return FALSE;
-    }
-
-#ifdef CLOCK_MONOTONIC
-    clock_gettime(CLOCK_MONOTONIC, &after_t);
-
-    crm_info("Compressed %d bytes into %d (ratio %d:1) in %dms",
-             length, *result_len, length / (*result_len),
-             (after_t.tv_sec - before_t.tv_sec) * 1000 + (after_t.tv_nsec -
-                                                          before_t.tv_nsec) / 1000000);
-#else
-    crm_info("Compressed %d bytes into %d (ratio %d:1)",
-             length, *result_len, length / (*result_len));
-#endif
-
-    *result = compressed;
-    return TRUE;
-}
-
 #ifdef HAVE_GNUTLS_GNUTLS_H
 void
 crm_gnutls_global_init(void)
@@ -2457,4 +2144,3 @@ crm_gnutls_global_init(void)
     gnutls_global_init();
 }
 #endif
-
